Backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
black==22.1.0
blue==0.9.1
cachetools==5.5.1
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.7
django-auditlog==3.0.0
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-ranged-response==0.2.0
django-rql==4.4.1
django-simple-captcha==0.6.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
flake8==4.0.1
idna==3.10
inflection==0.5.1
isort==6.0.1
kombu==5.5.4
lark-parser==0.11.0
lib-rql==2.0.2
mccabe==0.6.1
mypy_extensions==1.1.0
packaging==24.2
panda==0.3.1
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
redis==5.0.1
pandas==2.1.3
openpyxl==3.1.2
python-magic==0.4.27
